pyproject.toml
[tool.poetry]
name = "open-search-test"
version = "0.1.0"
description = ""
authors = ["Mouad Kommir <mouad.kommir@payticconnect.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic-config = "^0.3.0"
opensearch-py = "^2.8.0"
pydantic-settings = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
